requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.14.0
git+https://github.com/KichangKim/DeepDanbooru.git
h5py==3.8.0